+++ Backend/src/main/java/com/carelink/appointment/repository/DoctorAvailabilityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DoctorAvailabilityRepository extends JpaRepository<DoctorAvailability, Long> {
-    List<DoctorAvailability> findByDoctorId(Long doctorId);
+public interface DoctorAvailabilityRepository extends JpaRepository<DoctorAvailability, Integer> {
+    List<DoctorAvailability> findByDoctorId(Integer doctorId);
 }
